// app/(private)/area-administrativa/configuracion/models/import-patient.tsx
export type ImportPatient = {
  fechaFiliacion: Date | undefined;
  apellidoPaterno: string | undefined;
  apellidoMaterno: string | undefined;
  primerNombre: string | undefined;
  segundoNombre: string | undefined;
  fechaDeNacimiento: Date | undefined;
  lugarDeNacimiento: string | undefined;
  carnetDeIdentidad: string | undefined;
  direccion: string | undefined;
  telefono: string | undefined;
  celular: string | undefined;
  email: string | undefined;
  referidoPor: string | undefined;
  motivoDeConsulta: string | undefined;
  alergias: string | undefined;
};